change some attribute from Long to Double

+++ src/main/java/com/example/mathang/SuaMatHangServlet.java
@@ -148,8 +148,8 @@ public class SuaMatHangServlet extends HttpServlet {
         String name = toUTF8String(request.getParameter("productname"));
         Part part = request.getPart("img");
         System.out.println(part.getContentType());
-        Long retailPrice = Long.parseLong(request.getParameter("productretailprice"));
-        Long wholesalePrice = Long.parseLong(request.getParameter("productwholesaleprice"));
+        Double retailPrice = Double.parseDouble(request.getParameter("productretailprice"));
+        Double wholesalePrice = Double.parseDouble(request.getParameter("productwholesaleprice"));
         Integer unit = Integer.parseInt(request.getParameter("weightunit"));
         String calculateUnit = toUTF8String(request.getParameter("productunit"));
         Float weight = Float.parseFloat(request.getParameter("productweight"));
@@ -237,8 +237,8 @@ public class SuaMatHangServlet extends HttpServlet {
         return category;
     }
 
-    public MatHang buildMatHang(Long id, String code, String name, InputStream image, Long retailPrice,
-                                Long wholesalePrice, Integer unit, String calculateUnit,
+    public MatHang buildMatHang(Long id, String code, String name, InputStream image, Double retailPrice,
+                                Double wholesalePrice, Integer unit, String calculateUnit,
                                 Float weight, String description, Category category, String attribute) {
         MatHang mathang = MatHang.builder()
                 .id(id)
@@ -257,8 +257,8 @@ public class SuaMatHangServlet extends HttpServlet {
         return mathang;
     }
 
-    public MatHang buildMatHangWithoutImage(Long id, String code, String name, Long retailPrice,
-                                            Long wholesalePrice, Integer unit, String calculateUnit,
+    public MatHang buildMatHangWithoutImage(Long id, String code, String name, Double retailPrice,
+                                            Double wholesalePrice, Integer unit, String calculateUnit,
                                             Float weight, String description, Category category, String attribute) {
         MatHang mathang = MatHang.builder()
                 .id(id)
